+++ src/main/java/org/cactoos/list/ConcatIterator.java
@@ -61,14 +61,14 @@ public final class ConcatIterator<T> implements Iterator<T> {
 
     @Override
     public boolean hasNext() {
-        return new FilteredIterable<>(
+        return new FilteredOf<>(
             this.list, Iterator::hasNext
         ).iterator().hasNext();
     }
 
     @Override
     public T next() {
-        return new FilteredIterable<>(
+        return new FilteredOf<>(
             this.list, Iterator::hasNext
         ).iterator().next().next();
     }
