+++ src/main/java/org/cactoos/list/MapOf.java
@@ -118,7 +118,7 @@ public final class MapOf<X, Y> implements Map<X, Y> {
      */
     public <Z> MapOf(final Iterable<Z> list,
         final Func<Z, Map.Entry<X, Y>> entry) {
-        this(new MappedIterable<>(list, entry));
+        this(new MappedOf<>(list, entry));
     }
 
     /**
@@ -131,7 +131,7 @@ public final class MapOf<X, Y> implements Map<X, Y> {
      */
     public <Z> MapOf(final Map<X, Y> map, final Iterable<Z> list,
         final Func<Z, Map.Entry<X, Y>> entry) {
-        this(map, new MappedIterable<>(list, entry));
+        this(map, new MappedOf<>(list, entry));
     }
 
     /**
