+++ src/test/java/org/cactoos/list/EndlessOfTest.java
@@ -23,60 +23,25 @@
  */
 package org.cactoos.list;
 
-import java.util.Collections;
-import java.util.Comparator;
 import org.hamcrest.MatcherAssert;
 import org.hamcrest.Matchers;
 import org.junit.Test;
 
 /**
- * Test case for {@link SortedIterable}.
- *
- * @author Yegor Bugayenko (yegor256@gmail.com)
+ * Test Case for {@link EndlessOf}.
+ * @author Fabricio Cabral (fabriciofx@gmail.com)
  * @version $Id$
- * @since 0.7
+ * @since 0.11
  * @checkstyle JavadocMethodCheck (500 lines)
- * @checkstyle MagicNumberCheck (500 lines)
  */
-public final class SortedIterableTest {
+public final class EndlessOfTest {
 
     @Test
-    public void sortsAnArray() throws Exception {
+    public void endlessIterableTest() throws Exception {
         MatcherAssert.assertThat(
-            "Can't sort an iterable",
-            new SortedIterable<>(
-                new ArrayOf<>(
-                    3, 2, 10, 44, -6, 0
-                )
-            ),
-            Matchers.hasItems(-6, 0, 2, 3, 10, 44)
+            "Can't get unique endless iterable item",
+            new EndlessOf<>(1),
+            Matchers.hasItem(1)
         );
     }
-
-    @Test
-    @SuppressWarnings("PMD.AvoidDuplicateLiterals")
-    public void sortsAnArrayWithComparator() throws Exception {
-        MatcherAssert.assertThat(
-            "Can't sort an iterable with a comparator",
-            new SortedIterable<>(
-                Comparator.reverseOrder(),
-                new ArrayOf<>(
-                    "a", "c", "hello", "dude", "Friend"
-                )
-            ),
-            Matchers.hasItems("hello", "dude", "c", "a", "Friend")
-        );
-    }
-
-    @Test
-    public void sortsAnEmptyArray() throws Exception {
-        MatcherAssert.assertThat(
-            "Can't sort an empty iterable",
-            new SortedIterable<Integer>(
-                Collections.emptyList()
-            ),
-            Matchers.iterableWithSize(0)
-        );
-    }
-
 }
